light_analyzer/requirements.txt
Pillow>=9.0.0 # Use Pillow for PIL, good practice to version
numpy
scipy
matplotlib
